requirements.txt
accelerate>=0.23.0
torch>=2.1
torchvision
pillow
safetensors
tqdm
torchvision
transformers>=4.33
diffusers>=0.25.0
peft>=0.7.0
click>=8.1
pyyaml>=6.0.1
streamlit>=1.36